Update README.md
This program requires choice of which of several models to implement, and several edits of the code that follows to obtain the correct summary results and labeling on the graphs. Also, one can have either static or dynamic bifurcation parameter by editing the material right after the model choice line. If anyone has trouble implementing the program as desired please contact me.